fix(application): refactoring
