+++ src/components/features/volunteer-activities/VolunteerActivityPageClient.tsx
@@ -28,7 +28,8 @@ import { ErrorState } from '@/components/common/ui'
 import { useVolunteerActivityActions } from './hooks'
 
 const ApplicationModal = dynamic(
-	() => import('@/components/features/applications/ApplicationModal'),
+	() =>
+		import('@/components/features/applications/components/ApplicationModal'),
 	{
 		ssr: false,
 		loading: () => null,
